+++ lib/screens/word_list_screen.dart
@@ -1,42 +1,55 @@
-import 'package:flashcard/controllers/word_controller.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
+import '../controllers/word_controller.dart';
 import 'word_details_screen.dart';
 
 class WordListScreen extends StatelessWidget {
-  const WordListScreen({super.key});
+  final textEditingController = TextEditingController();
+
+  WordListScreen({super.key});
 
 // get controller to correct reflect words - add a test word, delete and check for update
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: const Text('Card list')),
+      appBar: AppBar(title: const Text('Word List')),
       body: Padding(
         padding: const EdgeInsets.all(8.0),
-        child: GetBuilder<WordController>(builder: (flashCardController) {
-          return ListView.builder(
-            itemCount: flashCardController.words.length,
-            itemBuilder: (context, index) => GestureDetector(
-              onTap: () async => await Get.to(() => WordDetailsScreen(title: 'Update Word', word: flashCardController.words[index])),
-              child: flashCardController.words.isEmpty
-                  ? Center(child: const Text('No words here. Add some words!')) // FIXME: this is not working
-                  : Column(
-                      children: [
-                        ListTile(
-                          title: Text(flashCardController.words[index].translation),
-                          subtitle: Text(flashCardController.words[index].hebrew),
-                        ),
-                        const Divider(
-                          thickness: 1,
-                          color: Colors.grey,
-                        ),
-                      ],
-                    ),
+        child: Column(
+          children: [
+            SearchBar(
+              controller: textEditingController,
+              onChanged: (value) => WordController.getOrPut.onSearchChange(value),
+              hintText: 'Search...',
             ),
-          );
-        }),
+            Expanded(
+              child: GetBuilder<WordController>(builder: (wordController) {
+                return ListView.builder(
+                  itemCount: wordController.filteredWords.length,
+                  itemBuilder: (context, index) => GestureDetector(
+                    onTap: () async => await Get.to(() => WordDetailsScreen(title: 'Update Word', word: wordController.filteredWords[index])),
+                    child: wordController.filteredWords.isEmpty
+                        ? Center(child: const Text('No words here. Add some words!')) // FIXME: this is not working
+                        : Column(
+                            children: [
+                              ListTile(
+                                title: Text(wordController.filteredWords[index].translation),
+                                subtitle: Text(wordController.filteredWords[index].hebrew),
+                              ),
+                              const Divider(
+                                thickness: 1,
+                                color: Colors.grey,
+                              ),
+                            ],
+                          ),
+                  ),
+                );
+              }),
+            ),
+          ],
+        ),
       ),
     );
   }