+++ lib/widgets/flash_card.dart
@@ -1,7 +1,7 @@
-import 'package:flashcard/controllers/word_controller.dart';
-import 'package:flashcard/models/constants.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+
+import '../controllers/word_controller.dart';
 
 class FlashCardWidget extends StatefulWidget {
   const FlashCardWidget({super.key});
@@ -10,23 +10,54 @@ class FlashCardWidget extends StatefulWidget {
   State<FlashCardWidget> createState() => _FlashCardWidgetState();
 }
 
-class _FlashCardWidgetState extends State<FlashCardWidget> {
-  bool _showBody = false;
+class _FlashCardWidgetState extends State<FlashCardWidget> with TickerProviderStateMixin {
+  late AnimationController animationController;
+  late Animation<double> animation;
+  AnimationStatus animationStatus = AnimationStatus.dismissed;
+  static const progressIndicatorDuration = 4;
+  static const opacityAnimationDuration = 2;
+  bool showTranslation = false;
+  bool enabledOnTap = true;
+  bool startedOpacityAnimation = false;
+
+  @override
+  void initState() {
+    _initAnimationController();
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    animationController.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
-    _delayShowingTranslation();
+    final showProgressIndicator = animationStatus == AnimationStatus.forward;
 
-    return GetBuilder<WordController>(builder: (flashCardController) {
+    if (!showProgressIndicator) {
+      _delayShowingTranslation();
+    }
+
+    return GetBuilder<WordController>(builder: (wordController) {
       return GestureDetector(
         onTap: () {
+          if (animationStatus == AnimationStatus.forward || !enabledOnTap) {
+            return;
+          }
+
           setState(() {
-            _showBody = false;
-            flashCardController.onTap();
+            wordController.onTap();
+            animationController.reset();
+            animationController.forward();
+            showTranslation = false;
+            enabledOnTap = false;
+            debugPrint('disabled');
           });
         },
-        onLongPress: flashCardController.onLongPress,
-        // need controller here to update word list after a deletion
+        onLongPress: wordController.onLongPress,
+        // TODO: need controller here to update word list after a deletion
         child: Card(
           elevation: 5,
           child: Padding(
@@ -35,31 +66,50 @@ class _FlashCardWidgetState extends State<FlashCardWidget> {
               mainAxisSize: MainAxisSize.min,
               children: [
                 Text(
-                  flashCardController.displayedWord.hebrew,
+                  wordController.displayedWord.hebrew,
                   style: Theme.of(context).textTheme.headlineLarge,
                 ),
-                AnimatedOpacity(
-                  opacity: _showBody ? 1.0 : 0.0,
-                  duration: _showBody
-                      ? const Duration(milliseconds: Constants.hideDuration)
-                      : Duration.zero, // TODO: display circular countdown timer before showing pronounciation/meaning
-                  child: Column(children: [
-                    Text(
-                      flashCardController.displayedWord.pronunciation,
-                      style: const TextStyle(fontSize: 16),
+                Stack(children: [
+                  // TODO: feature: scale by screenSize/cardSize -a Card doesn't have constrainst, so LayoutBuilder doesn't work here. At least check with GPT how a can get a sizebox to be a percentage of the current layout
+                  Opacity(
+                    opacity: showProgressIndicator ? 1 : 0.0,
+                    child: SizedBox(
+                      height: 48,
+                      width: 48,
+                      child: Center(
+                        child: SizedBox(
+                          height: 26,
+                          width: 26,
+                          child: CircularProgressIndicator(
+                            value: animation.value,
+                            valueColor: AlwaysStoppedAnimation<Color>(Colors.lightBlue.shade300),
+                            strokeWidth: 1,
+                          ),
+                        ),
+                      ),
                     ),
-                    Text(
-                      flashCardController.displayedWord.translation,
-                      style: const TextStyle(fontSize: 12, color: Colors.grey), // TODO: move to theme
-                    ),
-                    flashCardController.displayedWord.attributes.isNotEmpty
-                        ? Text(
-                            flashCardController.displayedWord.attributes,
-                            style: const TextStyle(fontSize: 12, fontStyle: FontStyle.italic, color: Colors.grey),
-                          )
-                        : SizedBox(),
-                  ]),
-                )
+                  ),
+                  if (!showProgressIndicator)
+                    AnimatedOpacity(
+                        duration: showTranslation ? Duration(seconds: opacityAnimationDuration) : Duration.zero,
+                        opacity: showTranslation ? 1.0 : 0.0,
+                        child: Column(children: [
+                          Text(
+                            wordController.displayedWord.pronunciation,
+                            style: Theme.of(context).textTheme.bodyMedium,
+                          ),
+                          Text(
+                            wordController.displayedWord.translation,
+                            style: Theme.of(context).textTheme.bodySmall!.copyWith(color: Colors.grey),
+                          ),
+                          wordController.displayedWord.attributes.isNotEmpty
+                              ? Text(
+                                  wordController.displayedWord.attributes,
+                                  style: Theme.of(context).textTheme.bodySmall!.copyWith(color: Colors.grey, fontStyle: FontStyle.italic),
+                                )
+                              : SizedBox(height: 12),
+                        ])),
+                ]),
               ],
             ),
           ),
@@ -68,13 +118,34 @@ class _FlashCardWidgetState extends State<FlashCardWidget> {
     });
   }
 
+  void _initAnimationController() {
+    animationController = AnimationController(
+      duration: const Duration(seconds: progressIndicatorDuration),
+      vsync: this,
+    );
+
+    animationController.addStatusListener((status) => setState(() => animationStatus = status));
+    animation = Tween<double>(begin: 0, end: 1).animate(animationController)..addListener(() => setState(() {}));
+    animationController.forward();
+  }
+
   _delayShowingTranslation() {
-    if (!_showBody) {
-      Future.delayed(const Duration(seconds: 2), () {
+    if (!showTranslation) {
+      const startOpacityAnimationDelayMs = 100;
+
+      Future.delayed(const Duration(milliseconds: startOpacityAnimationDelayMs), () {
         setState(() {
-          _showBody = true;
+          showTranslation = true;
         });
+      });
+
+      Future.delayed(Duration(milliseconds: startOpacityAnimationDelayMs + (opacityAnimationDuration * 1000)), () {
+        debugPrint('enabled');
+        setState(() => enabledOnTap = true);
       });
     }
   }
 }
+
+// TODO: change attributes to checkboxes: gender, plural, etc. Then color code the word. Will have to persist attribute settings too
+// add searching to WordList. Also, add ability to swop translation with hebrew word